agent/requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
MouseInfo==0.1.3
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
pytweening==1.2.0
requests==2.32.4
six==1.17.0
urllib3==2.5.0
